// src/redux/store/user/user.action-types.ts
export enum USER_ACTION_TYPES {
  SET_CURRENT_USER = 'user/SET_CURRENT_USER',
  CHECK_USER_SESSION = 'user/CHECK_USER_SESSION',
  GOOGLE_SIGN_IN_START = 'user/GOOGLE_SIGN_IN_START',
  CREATE_USER_EMAIL_PASSWORD_START = 'user/CREATE_USER_EMAIL_PASSWORD_START', // added by me for createAuthUserWithEmailAndPassword
  CREATE_USER_EMAIL_PASSWORD_SUCCESS = 'user/CREATE_USER_EMAIL_PASSWORD_SUCCESS', // added by me
  CREATE_USER_EMAIL_PASSWORD_FAILED = 'user/CREATE_USER_EMAIL_PASSWORD_FAILED', // added by me
  EMAIL_SIGN_IN_START = 'user/EMAIL_SIGN_IN_START',
  SIGN_IN_SUCCESS = 'user/SIGN_IN_SUCCESS',
  SIGN_IN_FAILED = 'user/SIGN_IN_FAILURE',
  SIGN_OUT_START = 'user/SIGN_OUT_START',
  SIGN_OUT_SUCCESS = 'user/SIGN_OUT_SUCCESS',
  SIGN_OUT_FAILED = 'user/SIGN_OUT_FAILED',
}

export default USER_ACTION_TYPES;