// src/utils/firebase/firebase.utils.types.ts
//the below type is for typeScript
//async function always returns a promise
export type ObjectToAdd = {
  title: string;
};
//again, defining a type for typeScript
export type AdditionalInformation = {
  displayName?: string;
};

//

export type UserData = {
  createdAt: Date;
  displayName: string;
  email: string;
};